requirements.txt
# Cryptography primitives (RIPEMD‑160 implementation)
pycryptodome>=3.20

# Human‑readable Base58 / Base58Check encoder‑decoder
base58>=2.1
